requirements.txt
tinyagi
streamlit
gitpython
python-dotenv
wikipedia-api
openai
llama-cpp-python
ollama
black
nltk
flask
wikipedia
emoji
PyYAML